Minor changes, removed comments
